django_react/requirements.txt
Django==4.2
gunicorn==20.1.0
dj-database-url==2.0.0
django-database-url==1.0.3
asgiref==3.6.0
psycopg2==2.9.6
sqlparse==0.4.4
djangorestframework==3.14.0
crispy-bootstrap5==0.7
django-crispy-forms==2.0